requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Report generation
reportlab>=4.0.0
Pillow>=10.0.0
jinja2>=3.1.0

# MCP (Model Context Protocol) integration
mcp>=0.1.0

# Data processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
